separate the controller in mutiple controller for correspond Solid method.
